requirements.txt
tweepy
requests
json
os
datetime
python-dotenv